refactor(product-selection): convert BakedGood coordinates from Double to BigDecimal
- Update BakedGood entity latitude/longitude fields to BigDecimal
- Update PublishBakedGoodsCommand coordinates to BigDecimal
- Update BakedGoodsPublishedEvent coordinates to BigDecimal
- Update BakedGoodDTO coordinates to BigDecimal
- Update PublishBakedGoodsRequestDTO coordinates to BigDecimal
- Update DevSeedController seed data to use BigDecimal literals
- Ensures consistency with ChosenLocation coordinate types

+++ product-selection-service/src/main/kotlin/org/pv293/kotlinseminar/productSelectionService/application/dto/PublishBakedGoodsRequestDTO.kt
@@ -13,7 +13,7 @@ data class PublishBakedGoodsRequestDTO(
     @field:Schema(example = "4.99")
     val price: BigDecimal,
     @field:Schema(example = "49.1951")
-    val latitude: Double,
+    val latitude: BigDecimal,
     @field:Schema(example = "16.6068")
-    val longitude: Double,
+    val longitude: BigDecimal,
 )
