+++ product-selection-service/src/main/kotlin/org/pv293/kotlinseminar/productSelectionService/events/impl/BakedGoodsPublishedEvent.kt
@@ -9,6 +9,6 @@ data class BakedGoodsPublishedEvent(
     val description: String?,
     val initialStock: Int,
     val price: BigDecimal,
-    val latitude: Double,
-    val longitude: Double,
+    val latitude: BigDecimal,
+    val longitude: BigDecimal,
 )
